+++ src/resources/views/pages/home.blade.php
@@ -15,8 +15,9 @@
                         <li>Added product variations. You can now add variations to each product. (issue #125)</li>
                         <li>Added reusable partial templates for Modal Window and Language selector tab. (issue #130)</li>
                         <li>Added shipping properties to products: Weight, Length, Width, Height, units. (issue #126)</li>
-                    </ol> 
-                    <h3>Version 0.3.2</h3>
+                    </ol>
+                    <hr>
+                    <h3>Version 0.3.2 and 0.2.2</h3>
                     <p>Code clean up, new features and UI improvements.</p>
                     <p>This update is generally v0.3.* compatible but there's a change with UserPricelist. Refer to UPGRADE.md for the upgrading instructions.</p>
                     <h4>Enhancement:</h4>
@@ -32,7 +33,8 @@
                     <ol>
                         <li>(legacy support) Purchase export to excel not working. Purchase has deprecated. Use Orders instead. (issue #115)</li>
                     </ol>
-                    <h3>Version 0.3.1</h3>
+                    <hr>
+                    <h3>Version 0.3.1 and 0.2.1</h3>
                     <p>New Bundle module and improvement to UI.</p>
                     <h4>New feature:</h4>
                     <ol>
@@ -54,12 +56,35 @@
                     <h4>Important:</h4>
                     <p>If you're upgrading from v0.2.0, please refer to the Upgrade Guide in README.md.</p>
                     <hr>
-                    <h3>Version 0.2.*</h3>
-                    <p>Focus on supporting Laravel 5.0 and making sure all models and controllers work.</p>
-                    <p>Version 0.2.0 is <strong>SOMEWHAT</strong> backward compatible.</p>
-                    <p>If you're upgrading from v0.1.*, please refer to the <a href="https://github.com/redooor/redminportal/blob/v0.2/UPGRADE.md">branch v0.2 Upgrade Guide</a>.</p>
-                    <p>Refer to <a href="https://github.com/redooor/redminportal/blob/v0.2/CHANGELOG.md">branch v0.2 CHANGELOG.md</a></p>
-
+                    <h3>Version 0.2.0</h3>
+                    <p>Focus on supporting Laravel 5.0.</p>
+                    <h4>New feature:</h4>
+                    <ol>
+                        <li>Support Laravel 5.0.</li>
+                        <li>Removed dependancy of Cartalyst/Sentry.</li>
+                        <li>Use Laravel's Authentication instead of Cartalyst/Sentry.</li>
+                        <li>Users and Groups tables should work with previous schema. (But need to add column remember_token in users table)</li>
+                        <li>Drop Discount module support from v0.2.0 onwards (issue #63)</li>
+                        <li>Changed UI to use [Redmaterials](http://redmaterials.redooor.com) theme.</li>
+                        <li>Removed Bootstrap from redminportal.css. It should be linked separately.</li>
+                        <li>Consistent UI across all views (issue #76).</li>
+                        <li>Add active column to Pricelists table (issue #78).</li>
+                        <li>Add UI to views for additional active column in Pricelists table (issue #78).</li>
+                        <li>Rearrange menu to split Physical product store and Digital product store (issue #52).</li>
+                        <li>Promotions, Products and Portfolios now allow translation (issue #9).</li>
+                        <li>Use Laravel authentication, removed Sentry dependency (issue #81).</li>
+                        <li>Allow multiple Category images (issue #79).</li>
+                        <li>Add Post Model and Controller (issue #73).</li>
+                        <li>Revamp Translation implementation for Category, Media, Module, Portfolio, Product and Promotion (issue #85).</li>
+                        <li>Add Page model and Controller (issue #74).</li>
+                        <li>Save orders for products (issue #51).</li>
+                        <li>Download orders in CSV format.</li>
+                    </ol>
+                    <h4>Bug fixes:</h4>
+                    <ol>
+                        <li>Translation to support any language (issue #83).</li>
+                    </ol>
+                    <hr>
                     <h3>Version 0.1.*</h3>
                     <p>Supports Laravel 4.2.</p>
                     <p>Refer to <a href="https://github.com/redooor/redminportal/blob/v0.1/README.md#change-log">branch v0.1 README.md Changelog</a></p>